requirements.txt
attrs==17.2.0
boto3==1.7.4
botocore==1.10.4
certifi==2018.1.18
chalice==1.1.1
chardet==3.0.4
click==6.6
docutils==0.14
domovoi==1.5.6
enum34==1.1.6
idna==2.6
jmespath==0.9.3
neo4j-driver==1.5.3
python-dateutil==2.6.1
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
typing==3.5.3.0
urllib3==1.22
